Add to Cart

+++ php/food-code.php
@@ -8,6 +8,7 @@ $catResult = mysqli_query($conn, $sql) or die (mysqli_error($conn));
 //Add Product
 if(isset($_POST['addFood']))
 {
+        $foodCode =$_POST['foodcode'];
         $foodName=$_POST['foodname'];
         $foodPrice=$_POST['foodprice'];
         $category=$_POST['category'];
@@ -15,13 +16,27 @@ if(isset($_POST['addFood']))
         $target = "./images/".basename($_FILES['image']['name']);   
         $image = $_FILES['image']['name'];
 
-        $insertFoodQuery = "INSERT INTO food (food_Name, food_Price, food_Category, Image) values ('$foodName', '$foodPrice', '$category', '$target')";
-        $insertFoodSqli = mysqli_query($conn, $insertFoodQuery) or die (mysqli_error($conn));
-        $reset = "ALTER TABLE food DROP No;ALTER TABLE food AUTO_INCREMENT = 1;ALTER TABLE food ADD No int UNSIGNED NOT NULL AUTO_INCREMENT PRIMARY KEY FIRST;";
-        $resetIncrement = mysqli_multi_query($conn, $reset);
+        // Check if Code exist
+        $checkifCodeExistQuery = "SELECT IFNULL( (SELECT food_Code FROM food WHERE food_Code ='$foodCode') ,'Null') As 'foodCode'";
+        $checkifCodeExistSqli = mysqli_query($conn, $checkifCodeExistQuery) or die (mysqli_error($conn));
+        $checkifCodeExist = mysqli_fetch_assoc($checkifCodeExistSqli);
+        $checkifCodeExist = $checkifCodeExist['foodCode'];
 
-        if (move_uploaded_file($_FILES['image']['tmp_name'], $target)) {
-
+        if($checkifCodeExist == "Null") 
+        {
+            $insertFoodQuery = "INSERT INTO food (food_Code, food_Name, food_Price, food_Category, Image) values ('$foodCode', '$foodName', '$foodPrice', '$category', '$target')";
+            $insertFoodSqli = mysqli_query($conn, $insertFoodQuery) or die (mysqli_error($conn));
+            $reset = "ALTER TABLE food DROP No;ALTER TABLE food AUTO_INCREMENT = 1;ALTER TABLE food ADD No int UNSIGNED NOT NULL AUTO_INCREMENT PRIMARY KEY FIRST;";
+            $resetIncrement = mysqli_multi_query($conn, $reset);
+    
+            if (move_uploaded_file($_FILES['image']['tmp_name'], $target)) {
+    
+            }
+            echo '<script>location.href="./foods.php"</script>';
+        }
+        else
+        {
+            echo '<script>alert("Food Code already exist! Please enter new Food Code.")</script>';
         }
 }
